代码提交

+++ common-config/src/main/java/com/qzx/user/entity/video/VoVideoInfoEntity.java
@@ -46,7 +46,7 @@ public class VoVideoInfoEntity implements Serializable {
     /**
      * 视频时长
      */
-    private Integer videoDuration;
+    private Long videoDuration;
 
     /**
      * 视频状态
